Move RecordFieldsNaming. Remove deprecated or unsafe utils classes.
