+++ snapi-client/src/main/scala/raw/client/rql2/truffle/Rql2CsvWriter.scala
@@ -39,7 +39,7 @@ import raw.compiler.rql2.source.{
 }
 import com.fasterxml.jackson.dataformat.csv.{CsvFactory, CsvSchema}
 import com.fasterxml.jackson.dataformat.csv.CsvGenerator.Feature.STRICT_CHECK_FOR_QUOTING
-import raw.utils.RecordFieldsNaming
+import raw.client.utils.RecordFieldsNaming
 
 import java.io.IOException
 import java.io.OutputStream
