+++ snapi-client/src/main/scala/raw/client/rql2/truffle/Rql2JsonWriter.scala
@@ -14,8 +14,8 @@ package raw.client.rql2.truffle
 
 import com.fasterxml.jackson.core.{JsonEncoding, JsonFactory, JsonParser}
 import org.graalvm.polyglot.Value
+import raw.client.utils.RecordFieldsNaming
 import raw.compiler.rql2.Rql2TypeUtils
-import raw.utils.RecordFieldsNaming
 import raw.compiler.rql2.source._
 
 import java.io.{IOException, OutputStream}
